requirements.txt
Django>=5.2.7
djangorestframework>=3.16.1
django-cors-headers>=4.9.0
djangorestframework-simplejwt>=5.5.1
psycopg2-binary>=2.9.10
python-dotenv>=1.1.1
whitenoise>=6.7.0
pytz>=2024.1
reportlab>=4.2.0
